adding yet another workaround to get ESP working again
now abs will be correctly included, which should fix problems on a lot of platforms (not just embedded boards). Now cmath will be included if available, to have support for std::abs. If not available there is a small implementation for it (it existed before but in the wrong place).