first iteration of unit generation
This prepares parts to generate the all of the units from templates. In the future all units will be generated from these templates. This allows more flexible enabling and disabling of features. Also since most of the units have the same code this removes a lot of redundancies and allows for an easier creation of new units